requirements.txt
# Data & Analysis
pandas>=2.0.0
requests>=2.31.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Web Scraping
beautifulsoup4>=4.12.0

# Configuration
python-dotenv>=1.0.0

# Jupyter Notebook
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
